Move token-related functions to the right place and consistify naming
